Plugins: Prefix plugin wrappers with NPN.
This is to avoid possible conflicts with library functions. It's necessary
on OS/2 where the _System calling convention in GCC suppresses C++ mangling
(which discards namespaces) so e.g. a wrapper called _write() conflicts
with a library function write() that gets an underscore in assembler.

This may become unnecessary later when _System is fixed in GCC (see
https://github.com/psmedley/gcc/issues/3) although it's not a good
practice to call functions like _myfun in general, since the underscore
is usually reserved by library functions.