Force CRLF for *.cmd files.
This is necessary for .cmd that are OS/2 REXX (otherwise REXX will fail with
an error due to garbage in file). It also will not hurt for Windows .cmd files.
Since all .cmd in the tree are either OS/2 or Windows, it is safe to use the
generic pattern for this repository.