[OS/2] libevent: Fix 100% CPU load when communicating process dies.
This looks like a bug (or at least a specific behavior) of select() from
the OS/2 TCP/IP stack. The 100% CPU load is caused by the fact that
nobody tells the other process about the peer death so that it could
remove the fd in question from selecting and it is just constantly failing
with EBADF in a tight loop.