[OS/2] Make sure all threads have FPU exception handler installed.
This handler is to work around the unexpected FPU CW reset that may
happen when calling some system DOS APIs or loading some 3rd party
DLLs. This handler used to be installed in Firefox 10 and earlier
but updating to FF17 and FF24 caused some code overlap so that it
didn't get installed on thread 1.

This change also moves EXCEPTQ installation from top-level main()
functions to library functions that perform setup of the Mozilla
Runtime for all derived processes.