[OS/2] chromium: Define OS_POSIX in header as well.
OS_POSIX is already defined on OS/2 (along with OS_OS2) in two places:
ipc/chromium/chromium-config.mozbuild and ipc/chromium/moz.build. But that
apparently doesn't cover cases when chromium headers are included
outside of this module. Poor decision. It should have been defined in
a single source (the header file) in the first place.