[OS/2] Disable deploying generic JS-based file picker dialog.
This dialog, when provided, would obscure the native implementation. Given that
the native implementation is preferred (and the JS-based version isn't even ported
to OS/2) this was not what the user would want.

Note that this problem would only show up in development builds of Firefox because
the packing stage already excludes the JS-based implementation on OS/2 (since it
is packed in only on GTK-based platforms).

Closes #108.