osfile: Close files on success in File.copy.
OS/2 takes the Posix path in the osfile component and there you cannot
delete or rename a file if it's already open. There may be other Posix
OSes behaving this way so it should be done right (like in other places).

This affects many things and in particular installing some addons. See
https://github.com/bitwiseworks/mozilla-os2/issues/113 for details.
Closes #113.