[OS/2] Enable FFMEG/LIBAV usage.
Note that this also includes some experimental changes to actually
enable the MP4 reader/decoder in FF as it looks like this code was
never properly tested upstream. More over, it looks like this part
was redone in around FF 35 so this commit is mostly for historical
purposes (and to know what to look at in a newer version).