[OS/2] widget: Implement GfxInfo stub.
This stub does nothing and just returns zero/null values (actually, OS/2 simply
re-uses the gonk stub that does just the same, to avoid 100% duplication).

The lack of this class would cause a request for a "@mozilla.org/gfx/info;1"
component to fail and in some cases it could crash due to acessing a
null object (like in WebGL code). Closes #152.