[OS/2] Remove OS/2-specific hacks for debug symbols creation.
Now it follows the same way as on other platforms, i.e. --enable-debug
or --enable-debug-symbols will cause HLL debug data to be linked into
all generated EXEs and DLLs isntead of keeping them in separate .dbg
files. Note that if both options are omitted, there will be still
some small amount of debug info in the binaries. You might want to
strip it with emxomfstrip or lxlite (or pass the `-s` option to
the compiler and linker flags when configuring).