[OS/2] Set BEGINLIBPATH and LIBPATHSTRICT in Firefox launcher.
This is to make sure that a private copy of XUL.DLL and other
Firefox support libraries belonging to a given release is loded.
It allows to have two different Firefox releases to be started
in parallel and also to start Firefox directly via firefox.exe
w/o any additional wrapper scripts and w/o even changing to a
directory containing Mozilla runtime files.