[OS/2] Use -O3 for JS code and -Os for the rest.
This is similar to other platforms like Linux. This set of options seems to provide an acceptable balance between the size and the speed.
This commit also deliberately disables generation of SSE/SSE2 instructions by default in -O3 -march=pentium4 builds which cause crashes due to a bug in GCC for OS/2 (see https://github.com/psmedley/gcc/issues/28 for details).